+++ client/src/Router.js
@@ -3,16 +3,16 @@ import HomePage from './pages/HomePage';
 import PostDetailPage from './pages/PostDetailPage';
 import PostPage from './pages/PostPage';
 import QuestionPage from './pages/QuestionPage';
-import LoginPage from './pages/LoginPage';
-import SignUpPage from './pages/SignUpPage';
+import LoginPage from './pages/signInUp/LoginPage';
+import SignUpPage from './pages/signInUp/SignUpPage';
 import SubscribePage from './pages/SubscribePage';
 import QuestionBoardPage from './pages/QuestionBoardPage';
 import FreeBoardPage from './pages/FreeBoardPage';
 import SuggestionBoardPage from './pages/SuggestionBoardPage';
-import DashBoardPage from './pages/DashBoardPage';
-import UserImgPage from './pages/UserImgPage';
-import UserNamePage from './pages/UserNamePage';
-import UserPasswordPage from './pages/UserPasswordPage';
+import DashBoardPage from './pages/mypage/DashBoardPage';
+import UserImgPage from './pages/mypage/UserImgPage';
+import UserNamePage from './pages/mypage/UserNamePage';
+import UserPasswordPage from './pages/mypage/UserPasswordPage';
 import AnswerPostPage from './pages/AnswerPostPage'
 import EditBoardPage from './pages/EditBoardPage';
 import EditQuestionPage from './pages/EditQuestionPage';
