+++ client/src/pages/mypage/UserNamePage.js
@@ -1,90 +1,14 @@
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import BasicButton from '../components/BasicButton';
-import BorderLayout from '../components/BorderLayout';
-import UserPassword from '../components/UserPassword';
-import axiosInstance from '../utils/axiosInstance';
+import BasicButton from '../../components/BasicButton';
+import BorderLayout from '../../components/BorderLayout';
+import UserName from '../../components/UserName';
 
-const UserPasswordPage = () => {
+const UserNamePage = () => {
   const themeState = useSelector((state) => state.themeSlice).theme;
-  const {userPicture} = useSelector((state) => state.userInfoSlice);
-
-  const [passwordValid, setPasswordValid] = useState(false);
-  const [rePasswordValid, setRePasswordValid] = useState(false);
-  const [passwordDesc, setPasswordDesc] = useState('');
-  const [rePasswordDesc, setRePasswordDesc] = useState('');
-  const [inputValue, setInputValue] = useState({
-    newPassword: '',
-    checkPassword: '',
-  });
-  const { newPassword, checkPassword } = inputValue;
-
-  const regNumber = /[0-9]/g;
-  const regString = /[a-zA-Z]/g;
-  const regSpecialCharacter =
-    /[\{\}\[\]\/?.,;:|\)*~`!^\-_+<>@\#$%&\\\=\(\'\"]/g;
-
-  const passwordValidation = () => {
-    if (
-      7 < newPassword.length &&
-      newPassword.length < 21 &&
-      regNumber.test(newPassword) &&
-      regString.test(newPassword) &&
-      regSpecialCharacter.test(newPassword)
-    ) {
-      setPasswordValid(true);
-      setPasswordDesc('');
-    } else {
-      setPasswordValid(false);
-      setPasswordDesc(
-        '8~20자 영문 대 소문자, 숫자, 특수문자를 포함하여야 합니다. '
-      );
-    }
-  };
-
-  const rePasswordValidation = () => {
-    if (checkPassword === newPassword) {
-      setRePasswordValid(true);
-      setRePasswordDesc('');
-    } else {
-      setRePasswordValid(false);
-      setRePasswordDesc('비밀번호가 일치하지 않습니다.');
-    }
-  };
-
-  const handleInput = (e) => {
-    const { name, value } = e.target;
-    setInputValue({
-      ...inputValue,
-      [name]: value,
-    });
-  };
-
-  const handleOnClick = async (e) => {
-    e.preventDefault();
-
-    if (!newPassword || !checkPassword) {
-      alert('새 비밀번호를 작성해주세요.');
-      return;
-    }
-
-    if (passwordValid & rePasswordValid) {
-      try {
-        axiosInstance
-          .patch('/my-page/patch', {
-            password: newPassword,
-          })
-          .then((res) => {
-            alert('비밀번호가 변경되었습니다.');
-          });
-      } catch (err) {
-        alert('비밀 번호 조건에 따라 작성해주세요.');
-      }
-    }
-  };
+  const {userPicture} = useSelector((state)=>state.userInfoSlice);
 
   return (
     <>
@@ -99,32 +23,12 @@ const UserPasswordPage = () => {
                   className='icon'
                 />
               </a>
-              <UserProfileImage src={userPicture} />
-              <form>
-                <UserPasswordInput
-                  className='newPassword'
-                  name='newPassword'
+              <UserProfileImage src={userPicture}/>
+              <div>
+                <UserNameInput
                   themeState={themeState}
-                  type='password'
-                  placeholder='새 비밀번호'
-                  onChange={handleInput}
-                  onKeyUp={passwordValidation}
+                  placeholder='변경할 닉네임을 입력하세요.'
                 />
-                <div className='warning'>
-                  <span>{passwordDesc}</span>
-                </div>
-                <UserPasswordInput
-                  className='checkPassword'
-                  name='checkPassword'
-                  themeState={themeState}
-                  type='password'
-                  placeholder='새 비밀번호 확인'
-                  onChange={handleInput}
-                  onKeyUp={rePasswordValidation}
-                />
-                <div className='warning'>
-                  <span>{rePasswordDesc}</span>
-                </div>
                 <BasicButton
                   themeState={themeState}
                   width='30%'
@@ -133,9 +37,8 @@ const UserPasswordPage = () => {
                   backGroundColor='var(--color-orange)'
                   fontSize='1.3rem'
                   text='변경하기'
-                  onClick={handleOnClick}
                 />
-              </form>
+              </div>
             </div>
             <div className='web'>
               <UserProfileImage src={userPicture} />
@@ -170,6 +73,7 @@ const UserPasswordPage = () => {
                   backGroundColor='var(--color-orange)'
                   fontSize='1.8rem'
                   text='닉네임 변경'
+                  selected
                 />
               </a>
               <a href='/userpassword'>
@@ -181,20 +85,12 @@ const UserPasswordPage = () => {
                   backGroundColor='var(--color-orange)'
                   fontSize='1.8rem'
                   text='비밀번호 변경'
-                  selected
                 />
               </a>
             </div>
           </LeftContent>
           <RightContent themeState={themeState}>
-            <UserPassword
-              handleInput={handleInput}
-              handleOnClick={handleOnClick}
-              passwordValidation={passwordValidation}
-              rePasswordValidation={rePasswordValidation}
-              passwordDesc={passwordDesc}
-              rePasswordDesc={rePasswordDesc}
-            />
+            <UserName />
           </RightContent>
         </Layout>
       </BorderLayout>
@@ -246,19 +142,12 @@ const LeftContent = styled.div`
         margin-bottom: 2rem;
       }
 
-      form {
+      div {
         display: flex;
         flex-direction: column;
+        justify-content: center;
         align-items: flex-end;
         height: 100%;
-
-        .warning {
-          position: relative;
-          bottom: 3rem;
-          display: flex;
-          justify-content: start;
-          width: 100%;
-        }
       }
     }
   }
@@ -302,7 +191,7 @@ const RightContent = styled.div`
   }
 `;
 
-const UserPasswordInput = styled.input`
+const UserNameInput = styled.input`
   width: 100%;
   height: 3rem;
   background-color: ${(props) =>
@@ -318,4 +207,4 @@ const UserPasswordInput = styled.input`
   }
 `;
 
-export default UserPasswordPage;
+export default UserNamePage;
